Improve wording
I love this idea.  There's tons of static developer blogs which could be ported from http with the integration of a markdown pre-processor. So many followup ideas!

Here's a few edits to smooth out the README English.  Words like 'simply', 'just', 'of course' and 'basically' are considered bad practice in tech docs where the reader could be junior or senior.  It's best to avoid them.